chore: remove deprecated depre_auth.py and auth_dependency.py files
Deleted the deprecated depre_auth.py and auth_dependency.py files, as they are no longer needed in the codebase.

- Removed depre_auth.py file, which was deprecated and no longer used.
- Deleted auth_dependency.py file, which is now obsolete and unnecessary.
